fix(design-tokens): rename utils
