+++ packages/design-tokens/design-token-previews/types/Token.ts
@@ -1,3 +1,1 @@
 export type Token = { value: string; comment?: string };
-
-export type TokenEntry = [string, Token];
