+++ packages/design-tokens/design-token-previews/types/TokenEntry.ts
@@ -1,0 +1,3 @@
+import { Token } from "./Token";
+
+export type TokenEntry = [string, Token];
