+++ packages/design-tokens/design-token-previews/types/TokenTable.ts
@@ -1,4 +1,4 @@
-import { TokenEntry } from "./Token";
+import { TokenEntry } from "./TokenEntry";
 
 export type Row = Array<JSX.Element | string>;
